feat(combine-logic): new logic implement
Add a new function that take a string and inside a
list, we are going to save the char of the generated
word only if it's the first and the last char or th
word.

Whereas if the char is in the middle of the word we
leave the blank space.

Reference: [b5a1dcb5]